updates for Render

+++ migrations/versions/20201120_150602_create_users_table.py.py
@@ -54,7 +54,7 @@ def upgrade():
     sa.Column('senderId', sa.Integer(), nullable=False),
     sa.Column('receiverId', sa.Integer(), nullable=False),
     sa.Column('createdAt', sa.DateTime(), nullable=False),
-    sa.Column('amount', sa.Integer(), nullable=False),
+    sa.Column('amount', sa.Float(), nullable=False),
     sa.Column('status', sa.String(), nullable=False),
     sa.PrimaryKeyConstraint('id'),
     sa.ForeignKeyConstraint(['senderId'], ['accounts.id']),
